Fixed project after directory mess up.
Had a problem with sorting out the directories of Laravel and accidently
broke the project when I tried to edit it. So I restarted the project
and copied the right code across. Since the last commit I have added a
create feature to allow the addition of providers for users to review.
This will require authorisation for admins.