Fixed error in pulling database information
Initially the provider individual pages were not pulling information
about the specific provider that the user was visiting which just left a
blank space where the title was supposed to be. I looked throughout the
routes file and saw that the slug of the title was not being pulled
through properly to fetch the correct data from the database. I simply
changed this in the index file.