Added new reviews page and routes
Variables not being passed through properly which is causing there to be
no results on the reviews page. So I decided to move the reviews to a
separate page as the route model binding was not working properly.
However I am still having issues with it and this will need researching.