Set up if statements which alter master page
If statements in the master page check whether a user is logged in and
displays them a welcome message depending on whether theyre logged in or
not. This uses the Auth token. Additionally I have also added in that
buttons which are not needed are removed when the user is logged in. For
example, a logged in user does not need access to sign up and login
pages.