Managed to pull house information
Finally managed to manipulate the code in order to attach a house to a
user. Thankfully this now works and a user can pull in house information
for a profile. The next step is to pull in other tenants which belong to
the same property. This will be displayed on a user's personal profile
page.