Added new profile features
Had some complications with the database so I decided to ask Elder
Studios for their advice on what would be the best way to approach it.
They notified me about using flash messages to notify users about
profile completion which I have taken in to account. So far I have added
an if statement which checks if the user has a house id linked to their
profile. If they do not it gives them a message on their profile. I have
also made the page appear more attractive with some modern looking
features as well as adding the message to notify them.