Fixed error with house id not working
Fixed error where house id wasnt working properly and the system thought
that a user did not have a house when they actually did. Next is to put
in a catch system where a user can not access the page to make another
house once they have already made a house.