Fix alert not allowing new users to view profile
Changes had to be made to the profile page coding as the code was trying
to get ahold of null data. This just needed to be changed to look for a
null house rather than a specific field of the house table. Changes also
made in accordance with usability and validation testing.